Error Checks
I added an instruction check to SET.
I changed the CORE_DUMP so that it outputs the last three things on the
stack. if there are less than three things it stopes when everything is
off the stack.
Seems to Run find under test programs. hw2test, crazycount, count10. I
modified hw2test to test various stack conditions when core_dump is
called.